+++ internal/model/job.go
@@ -12,14 +12,14 @@ const (
 	)
 
 type Job struct {
-	ID                int     `json:"id" valid:"int, optional"`
-	HireManagerId     int     `json:"hireManagerId,string" valid:"int, optional"`
+	ID                int64     `json:"id" valid:"int, optional"`
+	HireManagerId     int64   `json:"hireManagerId,string" valid:"int, optional"`
 	Title             string  `json:"title" valid:"utfletternum, required"`
 	Description       string  `json:"description"valid:"- , optional"`
 	Files             string  `json:"files" valid:"-"`
 	SpecialityId      int     `json:"specialityId,string" valid:"int, optional"`
 	ExperienceLevelId int     `json:"experienceLevelId,string" valid:"in(1|2|3)"`
-	PaymentAmount      float64 `json:"paymentAmount,string" valid:"float"`
+	PaymentAmount     float64 `json:"paymentAmount,string" valid:"float"`
 	Country           string  `json:"country" valid:"utfletternum, optional"`
 	City              string  `json:"city" valid:"utfletternum, optional"`
 	JobTypeId         int     `json:"jobTypeId,string" valid:"int, optional"`
